Fixed compilation warnings
Signed/unsigned mismatches.

+++ doomsday/engine/portable/src/sys_network.c
@@ -549,7 +549,7 @@ void N_SendDataBufferReliably(void *data, size_t size, nodeid_t destination)
                           (unsigned long) (size + 2), result) );
 #endif
 
-    if(result != size + 2)
+    if((size_t)result != size + 2)
     {
         perror("Socket error");
     }
